requirements.txt
Django==3.0.6
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
pytest==5.4.2
pytest-django==3.9.0
